Modularizes screen components
