+++ App.js
@@ -1,34 +1,9 @@
 import * as React from 'react';
-import { StatusBar } from 'expo-status-bar';
-import { Button, StyleSheet, Text, View } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-
-function QuestionsScreen({ navigation }) {
-  return (
-    <View style={styles.container}>
-      <Text>Question:</Text>
-      <StatusBar style="auto" />
-      <Button
-        title="See Results"
-        onPress={() => navigation.navigate('R1')}
-      />
-    </View>
-  );
-}
-
-function ResultsScreen({ navigation }) {
-  return (
-    <View style={styles.container}>
-      <Text>Result:</Text>
-      <StatusBar style="auto" />
-      <Button
-        title="Go Home"
-        onPress={() => navigation.navigate('Home')}
-      />
-    </View>
-  );
-}
+import { HomeScreen } from "./src/screens/Home/index.js";
+import { QuestionScreen } from "./src/screens/Question/index.js";
+import { ResultsScreen } from "./src/screens/Results/index.js";
 
 const Stack = createNativeStackNavigator();
 
@@ -37,7 +12,7 @@ function App() {
     <NavigationContainer>
       <Stack.Navigator>
         <Stack.Screen name="Home" component={HomeScreen} />
-        <Stack.Screen name="Q1" component={QuestionsScreen} />
+        <Stack.Screen name="Q1" component={QuestionScreen} />
         <Stack.Screen name="R1" component={ResultsScreen} />
       </Stack.Navigator>
     </NavigationContainer>
@@ -45,12 +20,3 @@ function App() {
 }
 
 export default App;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
