--- Components/HomeScreen.ts
@@ -1,16 +1,0 @@
-import * as React from 'react';
-import { StatusBar } from 'expo-status-bar';
-import { Button, StyleSheet, Text, View } from 'react-native';
-
-// function HomeScreen({ navigation }) {
-//     return (
-//       <View style={styles.container}>
-//         <Text>Welcome to Calvin Location Guesser!!!</Text>
-//         <StatusBar style="auto" />
-//         <Button
-//           title="Start Quiz"
-//           onPress={() => navigation.navigate('Q1')}
-//         />
-//       </View>
-//     );
-//   }
