+++ package.json
@@ -11,8 +11,6 @@
   "dependencies": {
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.0",
-    "@types/react": "^18.0.21",
-    "@types/react-native": "^0.70.4",
     "expo": "~46.0.9",
     "expo-status-bar": "~1.4.0",
     "react": "18.0.0",
@@ -21,7 +19,12 @@
     "react-native-screens": "~3.15.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9"
+    "@babel/core": "^7.12.9",
+    "@types/jest": "^29.1.0",
+    "@types/react": "^18.0.21",
+    "@types/react-native": "^0.70.4",
+    "@types/react-test-renderer": "^18.0.0",
+    "typescript": "^4.8.4"
   },
   "private": true
 }
